% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search trough the entire regrtest code base for a keyword}
\usage{
search(searchpattern)
}
\arguments{
\item{searchpattern}{The searchpattern to look for.}
}
\value{
Displays the documentation of all matching functions.
}
\description{
This function searches all the roxygen doc entries for
the given search pattern. It then displays the documentation.
The search is case insensitive, and grep patterns can be used.
}
\examples{
search("predictor centered")
}
